requirements.txt
aiohttp==3.6.2
aiosqlite3==0.3.0
appdirs==1.4.4
astroid==2.4.2
async-timeout==3.0.1
attrs==19.3.0
black==19.10b0
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
discord.py==1.5.1
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
pathspec==0.8.0
pycparser==2.20
pylint==2.5.3
PyNaCl==1.4.0
python-dotenv==0.14.0
regex==2020.6.8
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
websockets==8.1
wrapt==1.12.1
yarl==1.4.2
youtube-dlc==2020.10.31
